Move get_geoschem_level_metadata to benchmark_utils script
gcpy/benchmark/modules/benchmark_utils.py
- Moved get_geoschem_level_metadata here
- Also need to import pandas as pd

gcpy/benchmark/modules/benchmark_model_vs_obs.py
- Now import get_geoschem_level_metadata from benchmark_utils

CHANGELOG.md
- Updated accordingly
